+++ core/common/src/ipc/IpcWebSocket.ts
@@ -13,7 +13,6 @@ import { IpcListener, IpcSocket, IpcSocketBackend, IpcSocketFrontend, RemoveFunc
 /** @internal */
 export abstract class IpcWebSocketTransport {
   public abstract send(message: IpcWebSocketMessage): void;
-  public abstract listen(handler: (message: IpcWebSocketMessage) => void): void;
 }
 
 /** @internal */
@@ -37,16 +36,17 @@ export interface IpcWebSocketMessage {
 /** @internal */
 export abstract class IpcWebSocket implements IpcSocket {
   public static transport: IpcWebSocketTransport;
+  public static receivers: Set<(evt: Event, message: IpcWebSocketMessage) => void> = new Set();
 
   protected _channels = new Map<string, Set<IpcListener>>();
 
   public constructor() {
-    IpcWebSocket.transport.listen(async (m) => this.broadcast(m));
+    IpcWebSocket.receivers.add(async (e, m) => this.broadcast(e, m));
   }
 
   public abstract send(channel: string, ...data: any[]): void;
 
-  public receive(channel: string, listener: IpcListener): RemoveFunction {
+  public addListener(channel: string, listener: IpcListener): RemoveFunction {
     let listeners = this._channels.get(channel);
     if (!listeners) {
       listeners = new Set();
@@ -59,7 +59,11 @@ export abstract class IpcWebSocket implements IpcSocket {
     return () => listeners!.delete(listener);
   }
 
-  private async broadcast(message: IpcWebSocketMessage) {
+  public removeListener(channel: string, listener: IpcListener) {
+    this._channels.get(channel)?.delete(listener);
+  }
+
+  private async broadcast(evt: Event, message: IpcWebSocketMessage) {
     if (message.type !== IpcWebSocketMessageType.Send && message.type !== IpcWebSocketMessageType.Push)
       return;
 
@@ -72,7 +76,7 @@ export abstract class IpcWebSocket implements IpcSocket {
       arg = [];
 
     for (const handler of handlers)
-      handler({}, ...arg);
+      handler(evt, ...arg);
   }
 }
 
@@ -83,7 +87,7 @@ export class IpcWebSocketFrontend extends IpcWebSocket implements IpcSocketFront
 
   public constructor() {
     super();
-    IpcWebSocket.transport.listen(async (m) => this.dispatch(m));
+    IpcWebSocket.receivers.add(async (e, m) => this.dispatch(e, m));
     FrontendIpc.initialize(this);
   }
 
@@ -100,7 +104,7 @@ export class IpcWebSocketFrontend extends IpcWebSocket implements IpcSocketFront
     });
   }
 
-  private async dispatch(message: IpcWebSocketMessage) {
+  private async dispatch(_evt: Event, message: IpcWebSocketMessage) {
     if (message.type !== IpcWebSocketMessageType.Response || !message.response)
       return;
 
@@ -115,11 +119,11 @@ export class IpcWebSocketFrontend extends IpcWebSocket implements IpcSocketFront
 
 /** @internal */
 export class IpcWebSocketBackend extends IpcWebSocket implements IpcSocketBackend {
-  private _handlers = new Map<string, (methodName: string, ...args: any[]) => Promise<any>>();
+  private _handlers = new Map<string, (event: Event, methodName: string, ...args: any[]) => Promise<any>>();
 
   public constructor() {
     super();
-    IpcWebSocket.transport.listen(async (m) => this.dispatch(m));
+    IpcWebSocket.receivers.add(async (e, m) => this.dispatch(e, m));
     BackendIpc.initialize(this);
   }
 
@@ -127,7 +131,7 @@ export class IpcWebSocketBackend extends IpcWebSocket implements IpcSocketBacken
     IpcWebSocket.transport.send({ type: IpcWebSocketMessageType.Push, channel, data });
   }
 
-  public handle(channel: string, handler: (methodName: string, ...args: any[]) => Promise<any>): RemoveFunction {
+  public handle(channel: string, handler: (event: Event, methodName: string, ...args: any[]) => Promise<any>): RemoveFunction {
     this._handlers.set(channel, handler);
 
     return () => {
@@ -136,7 +140,7 @@ export class IpcWebSocketBackend extends IpcWebSocket implements IpcSocketBacken
     };
   }
 
-  private async dispatch(message: IpcWebSocketMessage) {
+  private async dispatch(_evt: Event, message: IpcWebSocketMessage) {
     if (message.type !== IpcWebSocketMessageType.Invoke || !message.method)
       return;
 
@@ -148,7 +152,7 @@ export class IpcWebSocketBackend extends IpcWebSocket implements IpcSocketBacken
     if (typeof (args) === "undefined")
       args = [];
 
-    const response = await handler(message.method, ...args);
+    const response = await handler({} as any, message.method, ...args);
 
     IpcWebSocket.transport.send({
       type: IpcWebSocketMessageType.Response,