+++ core/frontend/src/render/webgl/HiliteUniforms.ts
@@ -8,7 +8,7 @@
 
 import { Hilite } from "@bentley/imodeljs-common";
 import { FloatRgb } from "./FloatRGBA";
-import { UniformHandle } from "./Handle";
+import { UniformHandle } from "./UniformHandle";
 import { Matrix3 } from "./Matrix";
 import { desync, sync } from "./Sync";