+++ test-apps/display-test-app/src/backend/DtaElectronMain.ts
@@ -2,12 +2,11 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import { dialog, OpenDialogOptions } from "electron";
 import * as path from "path";
 import { assert } from "@bentley/bentleyjs-core";
 import { ElectronBackend, ElectronBackendOptions } from "@bentley/electron-manager/lib/ElectronBackend";
 import { IpcHandler } from "@bentley/imodeljs-common";
-import { DtaIpcInterface, DtaIpcKey } from "../common/DtaIpcInterface";
+import { dtaChannel, DtaIpcInterface } from "../common/DtaIpcInterface";
 import { getRpcInterfaces, initializeDtaBackend } from "./Backend";
 
 const getWindowSize = () => {
@@ -31,10 +30,9 @@ const getWindowSize = () => {
 };
 
 class DtaIpcImpl extends IpcHandler implements DtaIpcInterface {
-  public get channelName() { return DtaIpcKey.Channel; }
-  public async getVersion() { return DtaIpcKey.Version; }
-  public async openFile(options: OpenDialogOptions) {
-    return dialog.showOpenDialog(options);
+  public get channelName() { return dtaChannel; }
+  public async sayHello() {
+    return "Hello from backend";
   }
 }