+++ test-apps/display-test-app/src/common/DtaIpcInterface.ts
@@ -3,14 +3,8 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 
-import { IpcInterface } from "@bentley/imodeljs-common";
-import { OpenDialogOptions, OpenDialogReturnValue } from "electron";
+export const dtaChannel = "dta";
 
-export enum DtaIpcKey {
-  Channel = "dta",
-  Version = "1.0.2",
-}
-
-export interface DtaIpcInterface extends IpcInterface {
-  openFile: (options: OpenDialogOptions) => Promise<OpenDialogReturnValue>;
+export interface DtaIpcInterface {
+  sayHello: () => Promise<string>;
 }