+++ core/bentley/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @bentley/bentleyjs-core
 
-This log was last generated on Fri, 08 Jan 2021 18:34:03 GMT and should not be manually modified.
+This log was last generated on Thu, 28 Jan 2021 13:39:27 GMT and should not be manually modified.
+
+## 2.11.0
+Thu, 28 Jan 2021 13:39:27 GMT
+
+### Updates
+
+- make BeEvent.raiseEvent type safe
 
 ## 2.10.3
 Fri, 08 Jan 2021 18:34:03 GMT